Conflicts:
	ProjetEcom/WebContent/js/main.js
	ProjetEcom/WebContent/partials/header.html
	ProjetEcom/src/servlets/InscriptionClientServlet.java

+++ ProjetEcom/src/model/Commande.java
@@ -15,41 +15,37 @@ import javax.persistence.OneToMany;
 
 import com.google.gson.annotations.Expose;
 
-
 @Entity
 public class Commande {
 	@Id
-	@GeneratedValue(strategy=GenerationType.AUTO)
+	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Expose
-	private long id;
+	private long				id;
 	@Expose
-	private Date dateDeVente;
+	private Date				dateDeVente;
 	@Expose
-	private int prixTotal;
-	
-	
-	@OneToMany(mappedBy="laCommande", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-	private Collection<Vente> lesVentes;
-	
-	@ManyToOne 
-	@JoinColumn(name="Client_id") 
-	private Client leClient;
-	
-	@ManyToOne 
-	@JoinColumn(name="MoyenPaiement_id") 
-	private MoyenPaiement leMoyenDePaiement;
-	
-	
+	private float				prixTotal;
+
+	@OneToMany(mappedBy = "laCommande", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+	private Collection<Vente>	lesVentes;
+
+	@ManyToOne
+	@JoinColumn(name = "Client_id")
+	private Client				leClient;
+
+	@ManyToOne
+	@JoinColumn(name = "MoyenPaiement_id")
+	private MoyenPaiement		leMoyenDePaiement;
+
 	public Commande() {
 	}
 
-	public Commande(Date dateDeVente, int prixTotal) {
+	public Commande(Date dateDeVente, float prixTotal) {
 		super();
 		this.dateDeVente = dateDeVente;
 		this.prixTotal = prixTotal;
 	}
 
-	
 	public long getId() {
 		return id;
 	}
@@ -66,18 +62,17 @@ public class Commande {
 		this.dateDeVente = dateDeVente;
 	}
 
-	public int getPrixTotal() {
+	public float getPrixTotal() {
 		return prixTotal;
 	}
 
-	public void setPrixTotal(int prixTotal) {
+	public void setPrixTotal(float prixTotal) {
 		this.prixTotal = prixTotal;
 	}
 
 	public Collection<Vente> getLesVentes() {
 		return lesVentes;
 	}
-	
 
 	public void setLesVentes(Collection<Vente> lesVentes) {
 		this.lesVentes = lesVentes;
@@ -86,7 +81,6 @@ public class Commande {
 	public Client getLeClient() {
 		return leClient;
 	}
-	
 
 	public void setLeClient(Client leClient) {
 		this.leClient = leClient;
@@ -96,11 +90,8 @@ public class Commande {
 		return leMoyenDePaiement;
 	}
 
-
 	public void setLeMoyenDePaiement(MoyenPaiement moyenDePaiement) {
 		this.leMoyenDePaiement = moyenDePaiement;
 	}
-	
-	
 
 }
